CVSL-1030 move entity classes up a level
