+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/probation/CommunityApiClient.kt
@@ -4,7 +4,6 @@ import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.http.MediaType
 import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.client.WebClient
-import uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.probation.entity.User
 
 @Component
 class CommunityApiClient(@Qualifier("oauthCommunityApiClient") val communityApiClient: WebClient) {
