+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/probation/Manager.kt
@@ -1,0 +1,7 @@
+package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.probation
+
+class Manager(
+  val code: String,
+  val name: Name,
+  val team: Team
+)
