+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/probation/Name.kt
@@ -1,0 +1,6 @@
+package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.probation
+
+class Name(
+  val forename: String,
+  val surname: String
+)
