+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/probation/User.kt
@@ -1,0 +1,5 @@
+package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.probation
+
+data class User(
+  val teams: List<Team>
+)
