--- src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/probation/entity/Identifiers.kt
@@ -1,5 +1,0 @@
-package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.probation.entity
-
-class Identifiers(
-  val crn: String
-)
