--- src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/probation/entity/Manager.kt
@@ -1,7 +1,0 @@
-package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.probation.entity
-
-class Manager(
-  val code: String,
-  val name: Name,
-  val team: Team
-)
