--- src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/probation/entity/Name.kt
@@ -1,6 +1,0 @@
-package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.probation.entity
-
-class Name(
-  val forename: String,
-  val surname: String
-)
