--- src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/probation/entity/ProbationSearchResponse.kt
@@ -1,5 +1,0 @@
-package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.probation.entity
-
-data class ProbationSearchResponse(
-  val content: List<ProbationSearchResult>?
-)
