--- src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/probation/entity/User.kt
@@ -1,5 +1,0 @@
-package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.probation.entity
-
-data class User(
-  val teams: List<Team>
-)
